+++ buy-app/src/main/java/il/ac/technion/cs/sd/buy/app/Parser.java
@@ -29,7 +29,7 @@ public abstract class Parser {
             System.out.println("orderId: " + entry.getKey() + " userId: " + entry.getValue().getUserId() + " productID: "
                     + entry.getValue().getProductId() + " latestAmount: " + entry.getValue().getLatestAmount());
             System.out.print("amount history: ");
-            for(Integer amount : entry.getValue().getAmountHistory()){
+            for(Long amount : entry.getValue().getAmountHistory()){
                 System.out.print(amount + " ");
             }
             System.out.println("cancelled: " + entry.getValue().isCancelled());
